cmd/beep: address buffer underruns by configuring 50 ms period size
I have seen sound cards with a terrible default period size that is so small there's no chance of filling the buffer in time. We also want a minimal buffer size since I have seen that a too large buffer
means we have to generate a lot of sound for playback to start. This allows testing low latency as well.